+++ Frontend/src/assets/sprite.scss
@@ -1,15 +1,36 @@
-
-$alarm-name: "alarm";
-$alarm-x: 65px;
-$alarm-y: 107px;
-$alarm-offset-x: -65px;
-$alarm-offset-y: -107px;
-$alarm-width: 32px;
-$alarm-height: 32px;
-$alarm-total-width: 598px;
-$alarm-total-height: 161px;
-$alarm-image: "./sprite.png";
-$alarm: (65px, 107px, -65px, -107px, 32px, 32px, 598px, 161px, "@src/assets/sprite.png", "alarm", );
+$icon-alarm-name: "icon_alarm";
+$icon-alarm-x: 0px;
+$icon-alarm-y: 107px;
+$icon-alarm-offset-x: 0px;
+$icon-alarm-offset-y: -107px;
+$icon-alarm-width: 32px;
+$icon-alarm-height: 32px;
+$icon-alarm-total-width: 598px;
+$icon-alarm-total-height: 139px;
+$icon-alarm-image: "sprite.png";
+$icon-alarm: (0px, 107px, 0px, -107px, 32px, 32px, 598px, 139px, "/src/assets/sprite.png", "icon_alarm", );
+$icon-search-name: "icon_search";
+$icon-search-x: 37px;
+$icon-search-y: 107px;
+$icon-search-offset-x: -37px;
+$icon-search-offset-y: -107px;
+$icon-search-width: 32px;
+$icon-search-height: 32px;
+$icon-search-total-width: 598px;
+$icon-search-total-height: 139px;
+$icon-search-image: "sprite.png";
+$icon-search: (37px, 107px, -37px, -107px, 32px, 32px, 598px, 139px, "/src/assets/sprite.png", "icon_search", );
+$icon-user-name: "icon_user";
+$icon-user-x: 74px;
+$icon-user-y: 107px;
+$icon-user-offset-x: -74px;
+$icon-user-offset-y: -107px;
+$icon-user-width: 32px;
+$icon-user-height: 32px;
+$icon-user-total-width: 598px;
+$icon-user-total-height: 139px;
+$icon-user-image: "sprite.png";
+$icon-user: (74px, 107px, -74px, -107px, 32px, 32px, 598px, 139px, "/src/assets/sprite.png", "icon_user", );
 $programmerground-name: "programmerground";
 $programmerground-x: 0px;
 $programmerground-y: 0px;
@@ -18,36 +39,14 @@ $programmerground-offset-y: 0px;
 $programmerground-width: 598px;
 $programmerground-height: 102px;
 $programmerground-total-width: 598px;
-$programmerground-total-height: 161px;
-$programmerground-image: "@src/assets/sprite.png";
-$programmerground: (0px, 0px, 0px, 0px, 598px, 102px, 598px, 161px, "@src/assets/sprite.png", "programmerground", );
-$project-icon-name: "projectIcon";
-$project-icon-x: 102px;
-$project-icon-y: 107px;
-$project-icon-offset-x: -102px;
-$project-icon-offset-y: -107px;
-$project-icon-width: 24px;
-$project-icon-height: 24px;
-$project-icon-total-width: 598px;
-$project-icon-total-height: 161px;
-$project-icon-image: "@src/assets/sprite.png";
-$project-icon: (102px, 107px, -102px, -107px, 24px, 24px, 598px, 161px, "@src/assets/sprite.png", "projectIcon", );
-$search-image-name: "searchImage";
-$search-image-x: 0px;
-$search-image-y: 107px;
-$search-image-offset-x: 0px;
-$search-image-offset-y: -107px;
-$search-image-width: 60px;
-$search-image-height: 54px;
-$search-image-total-width: 598px;
-$search-image-total-height: 161px;
-$search-image-image: "@src/assets/sprite.png";
-$search-image: (0px, 107px, 0px, -107px, 60px, 54px, 598px, 161px, "@src/assets/sprite.png", "searchImage", );
+$programmerground-total-height: 139px;
+$programmerground-image: "sprite.png";
+$programmerground: (0px, 0px, 0px, 0px, 598px, 102px, 598px, 139px, "/src/assets/sprite.png", "programmerground", );
 $spritesheet-width: 598px;
-$spritesheet-height: 161px;
-$spritesheet-image: "@src/assets/sprite.png";
-$spritesheet-sprites: ($alarm, $programmerground, $project-icon, $search-image, );
-$spritesheet: (598px, 161px, "@src/assets/sprite.png", $spritesheet-sprites, );
+$spritesheet-height: 139px;
+$spritesheet-image: "/src/assets/sprite.png";
+$spritesheet-sprites: ($icon-alarm, $icon-search, $icon-user, $programmerground, );
+$spritesheet: (598px, 139px, "/src/assets/sprite.png", $spritesheet-sprites, );
 
 @mixin sprite-width($sprite) {
   width: nth($sprite, 5);
@@ -74,7 +73,6 @@ $spritesheet: (598px, 161px, "@src/assets/sprite.png", $spritesheet-sprites, );
   @include sprite-width($sprite);
   @include sprite-height($sprite);
 }
-
 
 @mixin sprites($sprites) {
   @each $sprite in $sprites {
