+++ Frontend/src/components/header/headerImage.scss
@@ -1,6 +1,9 @@
 @import '@src/assets/sprite.scss';
 
-.projection_icon{
-  @include sprite($project-icon);
+.user_icon {
+  @include sprite($icon-user);
 }
 
+.alarm_icon {
+  @include sprite($icon-alarm);
+}
