+++ Frontend/src/pages/PlayGroundPage/index.tsx
@@ -8,7 +8,6 @@
 import React, { useState, useEffect, useRef } from 'react';
 import Header from '@src/components/header';
 import SearchBar from '@src/components/searchBar';
-import Bone from '@src/components/Common/bone';
 import { useDispatch, useSelector } from 'react-redux';
 import PlaygroundContent from '@src/components/playgroundContent';
 import { getAllPlaygrounds, getOnePlayground } from '@src/lib/axios/playground';
@@ -74,13 +73,7 @@ const PlayGroundPage = () => {
 			<Header />
 
 			<StyledComponent.mainContainer>
-				<StyledComponent.SearchContainer>
-					<SearchBar />
-					{/* <StyledComponent.ModeContainer>
-						<Bone />
-					</StyledComponent.ModeContainer> */}
-					<StyledComponent.CreateLink>방 생성</StyledComponent.CreateLink>
-				</StyledComponent.SearchContainer>
+				<StyledComponent.SearchContainer />
 				<StyledComponent.PlayGroundContainer>
 					{result.map((v) => {
 						return (
